app_tf_server/requirements.txt
flask
numpy
Pillow
werkzeug
flask-mysql
requests
scipy
pandas